functioning live API server with generator
