+++ src/classifai/generators/__init__.py
@@ -2,6 +2,6 @@ from .base import GeneratorBase
 from .gcp import RagAgent_GCP
 
 __all__ = [
-    "RagAgent_GCP",
     "GeneratorBase",
+    "RagAgent_GCP",
 ]
