+++ src/classifai/servers/main.py
@@ -1,31 +1,28 @@
 # pylint: disable=C0301
-"""This module provides functionality for creating a start a restAPI service which
+"""This module provides functionality for creating a restAPI service which
 allows a user to call the search methods of different VectorStore objects, from
-an api-endpoint.
+designated API endpoints.
 
-These functions interact with the ClassifAI PackageIndexer modules
+These functions interact with the ClassifAI Indexer module
 VectorStore objects, such that their embed and search methods are exposed on
 restAPI endpoints, in a FastAPI restAPI service started with these functions.
 """
 
 import logging
-from typing import Annotated
 
 import uvicorn
 
 # New imports
-from fastapi import FastAPI, Query
+from fastapi import FastAPI
 from fastapi.responses import RedirectResponse
 
-from .pydantic_models import (
-    ClassifaiData,
-    EmbeddingsList,
-    EmbeddingsResponseBody,
-    ResultsResponseBody,
-    RevClassifaiData,
-    RevResultsResponseBody,
-    convert_dataframe_to_pydantic_response,
-    convert_dataframe_to_reverse_search_pydantic_response,
+from ..generators.base import GeneratorBase
+from ..indexers import VectorStore
+from .endpoint_functions import (
+    create_embedding_endpoint,
+    create_rag_endpoint,
+    create_reverse_search_endpoint,
+    create_search_endpoint,
 )
 
 
@@ -45,116 +42,23 @@ def start_api(vector_stores, endpoint_names, port=8000):
     """
     logging.info("Starting ClassifAI API")
 
-    endpoint_index_map = {x: i for i, x in enumerate(endpoint_names)}
-
     app = FastAPI()
-
-    def create_embedding_endpoint(app, endpoint_name, vector_store):
-        """Create and register an embedding endpoint for a specific vector store.
-
-        Args:
-            app (FastAPI): The FastAPI application instance.
-            endpoint_name (str): The name of the endpoint to be created.
-            vector_store: The vector store object responsible for generating embeddings.
-
-        The created endpoint accepts POST requests with input data, generates embeddings
-        for the provided documents, and returns the results in a structured format.
-        """
-
-        @app.post(f"/{endpoint_name}/embed", description=f"{endpoint_name} embedding endpoint")
-        async def embedding_endpoint(data: ClassifaiData) -> EmbeddingsResponseBody:
-            input_ids = [x.id for x in data.entries]
-            documents = [x.description for x in data.entries]
-
-            embeddings = vector_store.embed(documents)
-
-            returnable = []
-            for idx, desc, embed in zip(input_ids, documents, embeddings, strict=True):
-                returnable.append(
-                    EmbeddingsList(
-                        idx=idx,
-                        description=desc,
-                        embedding=embed.tolist(),
-                    )
-                )
-            return EmbeddingsResponseBody(data=returnable)
-
-    def create_search_endpoint(app, endpoint_name, vector_store):
-        """Create and register a search endpoint for a specific vector store.
-
-        Args:
-            app (FastAPI): The FastAPI application instance.
-            endpoint_name (str): The name of the endpoint to be created.
-            vector_store: The vector store object responsible for performing search operations.
-
-        The created endpoint accepts POST requests with input data and a query parameter
-        specifying the number of results to return. It performs a search operation using
-        the vector store and returns the results in a structured format.
-        """
-
-        @app.post(f"/{endpoint_name}/search", description=f"{endpoint_name} search endpoint")
-        async def search_endpoint(
-            data: ClassifaiData,
-            n_results: Annotated[
-                int,
-                Query(
-                    description="The number of knowledgebase results to return per input query.",
-                    ge=1,  # Ensure at least one result is returned
-                ),
-            ] = 10,
-        ) -> ResultsResponseBody:
-            input_ids = [x.id for x in data.entries]
-            queries = [x.description for x in data.entries]
-
-            query_result = vector_store.search(query=queries, ids=input_ids, n_results=n_results)
-            ##post processing of the pandas dataframe
-            formatted_result = convert_dataframe_to_pydantic_response(
-                df=query_result,
-                meta_data=vector_stores[endpoint_index_map[endpoint_name]].meta_data,
-            )
-
-            return formatted_result
-
-    def create_reverse_search_endpoint(app, endpoint_name, vector_store):
-        """Create and register a reverse_search endpoint for a specific vector store.
-
-        Args:
-            app (FastAPI): The FastAPI application instance.
-            endpoint_name (str): The name of the endpoint to be created.
-            vector_store: The vector store object responsible for performing search operations.
-
-        The created endpoint accepts POST requests with input data and a query parameter
-        specifying the number of results to return. It performs a reverse search operation using
-        the vector store and returns the results in a structured format.
-        """
-
-        @app.post(f"/{endpoint_name}/reverse_search", description=f"{endpoint_name} reverse query endpoint")
-        def reverse_search_endpoint(
-            data: RevClassifaiData,
-            n_results: Annotated[
-                int,
-                Query(
-                    description="The max number of results to return.",
-                ),
-            ] = 100,
-        ) -> RevResultsResponseBody:
-            input_ids = [x.id for x in data.entries]
-            queries = [x.code for x in data.entries]
-
-            reverse_query_result = vector_store.reverse_search(query=queries, ids=input_ids, n_results=n_results)
-
-            formatted_result = convert_dataframe_to_reverse_search_pydantic_response(
-                df=reverse_query_result,
-                meta_data=vector_stores[endpoint_index_map[endpoint_name]].meta_data,
-                ids=input_ids,
-            )
-            return formatted_result
 
     for endpoint_name, vector_store in zip(endpoint_names, vector_stores, strict=True):
         logging.info("Registering endpoints for: %s", endpoint_name)
-        create_embedding_endpoint(app, endpoint_name, vector_store)
-        create_search_endpoint(app, endpoint_name, vector_store)
-        create_reverse_search_endpoint(app, endpoint_name, vector_store)
+
+        # Check if the object is an instance of a GeneratorBase subclass or VectorStore and load the corresponding endpoints
+        if isinstance(vector_store, GeneratorBase):
+            create_rag_endpoint(app, endpoint_name, vector_store)
+        elif isinstance(vector_store, VectorStore):
+            create_embedding_endpoint(app, endpoint_name, vector_store)
+            create_search_endpoint(app, endpoint_name, vector_store)
+            create_reverse_search_endpoint(app, endpoint_name, vector_store)
+        else:
+            raise TypeError(
+                f"Unsupported object type for endpoint '{endpoint_name}': {type(vector_store)}. "
+                "Expected VectorStore or GeneratorBase instance."
+            )
 
     @app.get("/", description="UI accessibility")
     def docs():
